+++ cw-shapes/src/main/java/edu/uob/TwoDimensionalShape.java
@@ -22,5 +22,5 @@ public abstract class TwoDimensionalShape {
 
   public abstract double calculateArea();
 
-  public abstract int calculatePerimeterLength();
+  public abstract long calculatePerimeterLength();
 }
